requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
atomicwrites==1.4.0
attrs==20.2.0
click==7.1.2
colorama==0.4.4
driver==0.1.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
